Optimize more scripts

+++ catscripts/Entertainment/Music/Lyrics/Lyrics.sh
@@ -1,14 +1,31 @@
 #!/bin/bash
 
-egrep -i 'lyrics|lyric\b' newpages.txt >> Lyrics.txt
+KEYWORDS_LYRICS="lyrics|lyric\b|words(| )to(| )the(| )song"
+KEYWORDS_LYRICS_ALL="$KEYWORDS_LYRICS"
 
-LYRICS=`stat --print=%s Lyrics.txt`
-
-if [ $LYRICS -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Lyrics.txt"
-  export CATNAME="Lyrics"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Lyrics\n"
+  fi
 
-rm Lyrics.txt
+  LYRICS=`egrep -i "$KEYWORDS_LYRICS" newpages.txt`
+
+  if [ "$LYRICS" != "" ];
+  then
+    printf "$LYRICS" > Lyrics.txt
+    export CATFILE="Lyrics.txt"
+    export CATNAME="Lyrics"
+    $CATEGORIZE
+    rm Lyrics.txt
+    unset LYRICS
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Lyrics\n"
+  fi
+
+fi
