+++ catscripts/Government/Countries/Latvia/Latvia.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'Latvia' newpages.txt >> Latvia.txt
+KEYWORDS_LATVIA="Latvia"
 
-LATVIA=`stat --print=%s Latvia.txt`
-
-if [ $LATVIA -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Latvia.txt"
-  export CATNAME="Latvia"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Latvia\n"
+  fi
 
-rm Latvia.txt
+  LATVIA=`egrep -i "$KEYWORDS_LATVIA" newpages.txt`
+
+  if [ "$LATVIA" != "" ];
+  then
+    printf "$LATVIA" > Latvia.txt
+    export CATFILE="Latvia.txt"
+    export CATNAME="Latvia"
+    $CATEGORIZE
+    rm Latvia.txt
+    unset LATVIA
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Latvia\n"
+  fi
+
+fi
