+++ catscripts/Lifestyle/Health/Diseases/Asthma/Asthma.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'asthma' newpages.txt > Asthma.txt
+KEYWORDS_ASTHMA="Asthma"
 
-ASTHMA=`stat --print=%s Asthma.txt`
-
-if [ $ASTHMA -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Asthma.txt"
-  export CATNAME="Asthma"
-  $CATEGORIZE
-fi
 
-rm Asthma.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Asthma\n"
+  fi
+
+  ASTHMA=`egrep -i "$KEYWORDS_ASTHMA" newpages.txt`
+
+  if [ "$ASTHMA" != "" ];
+  then
+    printf "$ASTHMA" > Asthma.txt
+    export CATFILE="Asthma.txt"
+    export CATNAME="Asthma"
+    $CATEGORIZE
+    rm Asthma.txt
+    unset ASTHMA
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Asthma\n"
+  fi
+
+fi
