+++ catscripts/Maintenance/Pornography.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'porn|hentai|lolicon|futanari|sex.+videos|videos.+sex' newpages.txt >> Pornography.txt
+KEYWORDS_PORNOGRAPHY="porn|hentai|lolicon|futanari|sex.+videos|videos.+sex"
 
-PORN=`stat --print=%s Pornography.txt`
-
-if [ $PORN -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Pornography.txt"
-  export CATNAME="Pornography"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Pornography maintenance script\n"
+  fi
 
-rm Pornography.txt
+  PORN=`egrep -i "$KEYWORDS_PORNOGRAPHY" newpages.txt`
+
+  if [ "$PORN" != "" ];
+  then
+    printf "$PORN" > Pornography.txt
+    export CATFILE="Pornography.txt"
+    export CATNAME="Pornography"
+    $CATEGORIZE
+    rm Pornography.txt
+    unset PORN
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Pornography maintenance script\n"
+  fi
+
+fi
